fix: Merge한 이후, package나 Type오류 가 나타나는 현상 수정

+++ src/main/java/com/shop/coffee/order/repository/OrderRepository.java
@@ -1,5 +1,4 @@
 package com.shop.coffee.order.repository;
-
 import com.shop.coffee.order.OrderStatus;
 import com.shop.coffee.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 
     // 주문 상태에 따라 주문 조회
     List<Order> findByOrderStatus(OrderStatus orderStatus);
-    
+
     //전체 주문 조회
     List<Order> findAll();
     // 주문 생성 시간 내림차순 정렬
